src/pkg/exp/html/testdata/webkit/tests26.dat

#data
<!DOCTYPE html><body><a href='#1'><nobr>1<nobr></a><br><a href='#2'><nobr>2<nobr></a><br><a href='#3'><nobr>3<nobr></a>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <a>
|       href="#1"
|       <nobr>
|         "1"
|       <nobr>
|     <nobr>
|       <br>
|       <a>
|         href="#2"
|     <a>
|       href="#2"
|       <nobr>
|         "2"
|       <nobr>
|     <nobr>
|       <br>
|       <a>
|         href="#3"
|     <a>
|       href="#3"
|       <nobr>
|         "3"
|       <nobr>

#data
<!DOCTYPE html><body><b><nobr>1<nobr></b><i><nobr>2<nobr></i>3
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|       <nobr>
|     <nobr>
|       <i>
|     <i>
|       <nobr>
|         "2"
|       <nobr>
|     <nobr>
|       "3"

#data
<!DOCTYPE html><body><b><nobr>1<table><nobr></b><i><nobr>2<nobr></i>3
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|         <nobr>
|           <i>
|         <i>
|           <nobr>
|             "2"
|           <nobr>
|         <nobr>
|           "3"
|         <table>

#data
<!DOCTYPE html><body><b><nobr>1<table><tr><td><nobr></b><i><nobr>2<nobr></i>3
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|         <table>
|           <tbody>
|             <tr>
|               <td>
|                 <nobr>
|                   <i>
|                 <i>
|                   <nobr>
|                     "2"
|                   <nobr>
|                 <nobr>
|                   "3"

#data
<!DOCTYPE html><body><b><nobr>1<div><nobr></b><i><nobr>2<nobr></i>3
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|     <div>
|       <b>
|         <nobr>
|         <nobr>
|       <nobr>
|         <i>
|       <i>
|         <nobr>
|           "2"
|         <nobr>
|       <nobr>
|         "3"

#data
<!DOCTYPE html><body><b><nobr>1<nobr></b><div><i><nobr>2<nobr></i>3
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|       <nobr>
|     <div>
|       <nobr>
|         <i>
|       <i>
|         <nobr>
|           "2"
|         <nobr>
|       <nobr>
|         "3"

#data
<!DOCTYPE html><body><b><nobr>1<nobr><ins></b><i><nobr>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|       <nobr>
|         <ins>
|     <nobr>
|       <i>
|     <i>
|       <nobr>

#data
<!DOCTYPE html><body><b><nobr>1<ins><nobr></b><i>2
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       <nobr>
|         "1"
|         <ins>
|       <nobr>
|     <nobr>
|       <i>
|         "2"

#data
<!DOCTYPE html><body><b>1<nobr></b><i><nobr>2</i>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <b>
|       "1"
|       <nobr>
|     <nobr>
|       <i>
|     <i>
|       <nobr>
|         "2"
